feat(auto-format): run auto-format in subdirectories

+++ app/src/main/java/com/asfoundation/wallet/ui/camera/CameraSource.java
@@ -756,7 +756,7 @@ public class CameraSource {
    */
   private byte[] createPreviewBuffer(Size previewSize) {
     int bitsPerPixel = ImageFormat.getBitsPerPixel(ImageFormat.NV21);
-    long sizeInBits = previewSize.getHeight() * previewSize.getWidth() * bitsPerPixel;
+    long sizeInBits = (long) previewSize.getHeight() * previewSize.getWidth() * bitsPerPixel;
     int bufferSize = (int) Math.ceil(sizeInBits / 8.0d) + 1;
 
     //
